refactor(format): rename ndjson json_each_row
