+++ query/src/formats/mod.rs
@@ -19,7 +19,7 @@ mod format_parquet;
 pub mod format_tsv;
 pub mod output_format;
 pub mod output_format_csv;
-mod output_format_ndjson;
+mod output_format_json_each_row;
 mod output_format_parquet;
 mod output_format_values;
