+++ query/src/formats/output_format_json_each_row.rs
